fix(release): trigger only on v* tag push; drop unworkable UI-draft path
A draft release never creates a git tag, so no push/create event
fires and GitHub fires no release event for draft creation — the
`create`-triggered "UI draft" path could never run. Make `v*` tag
push the sole trigger, remove the trigger-guard job and the
check-release-trigger command it invoked, simplify the env/
concurrency/job conditionals, and always publish the draft as the
final atomic step. Relocate the shared releaseLookupError into
publishrelease.go. Update release.md and release-tooling.md.
