+++ internal/release/publishrelease_test.go
@@ -16,6 +16,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+type roundTripFunc func(*http.Request) (*http.Response, error)
+
+func (f roundTripFunc) RoundTrip(req *http.Request) (*http.Response, error) {
+	return f(req)
+}
+
 func TestPublishReleaseFlipsDraftToPublished(t *testing.T) {
 	var (
 		mu      sync.Mutex
